feat: Sync chats deletion across devices
Motivation: Now broadcast lists creation is synced across devices. Groups creation, in some means, too -- on a group promotion by a first message. So, to make this looking more feature-complete, sync chats deletion too.

An approach is to replay a user action on synchronised devices because the chat deletion is a complex procedure affecting several tables, self-chat, emitting the corresponding event etc.

Also this commit makes all deletions to happen in a transaction, to be on a safe side.